make i3lock black after 2100

+++ i3/setup/scripts/lock.sh
@@ -10,8 +10,13 @@
 #convert "$tmpbg" "$icon" -gravity center -composite -matte "$tmpbg"
 #i3lock -i "$tmpbg"
 
-rm /tmp/screenshot.png
-scrot /tmp/screenshot.png
-convert /tmp/screenshot.png -blur 0x5 /tmp/screenshotblur.png
+if [ $(date +"%H") -ge 21 ]
+then
+  i3lock -c 000000
+else
+  rm /tmp/screenshot.png
+  scrot /tmp/screenshot.png
+  convert /tmp/screenshot.png -blur 0x5 /tmp/screenshotblur.png
 
-i3lock -i /tmp/screenshotblur.png
+  i3lock -i /tmp/screenshotblur.png
+fi
